midway through refactoring parallel lists to nested dictionary
