--- EnlistedPrefixes.csv
@@ -1,3 +1,0 @@
-K,Instructor
-Q,Evaluator
-T,Instructor
