--- OfficerPrefixes.csv
@@ -1,4 +1,0 @@
-K,Instructor
-Q,Evaluator
-T,Instructor
-W,Weapons Officer
